FHT: create hook to check if eligible for hosting trial
