+++ client/blocks/importer/wordpress/import-everything/pre-migration/upgrade-plan.tsx
@@ -8,7 +8,7 @@ import { useTranslate } from 'i18n-calypso';
 import React, { useRef, useState, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { convertToFriendlyWebsiteName } from 'calypso/blocks/import/util';
-import useCheckEligibilityMigrationTrialPlan from 'calypso/data/plans/use-check-eligibility-migration-trial-plan';
+import { useCheckMigrationTrialPlanEligibility } from 'calypso/data/plans/use-check-eligibility-migration-trial-plan';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import ConfirmUpgradePlan from './../confirm-upgrade-plan';
 import type { URL } from 'calypso/types';
@@ -38,7 +38,7 @@ export const PreMigrationUpgradePlan: React.FunctionComponent< Props > = ( props
 		isBusy,
 		migrationTrackingProps,
 	} = props;
-	const { data: migrationTrialEligibility } = useCheckEligibilityMigrationTrialPlan(
+	const { data: migrationTrialEligibility } = useCheckMigrationTrialPlanEligibility(
 		targetSite.ID
 	);
 	const isEligibleForTrialPlan = migrationTrialEligibility?.eligible;
