+++ client/landing/stepper/declarative-flow/internals/steps-repository/trial-acknowledge/migration-trial-acknowledge.tsx
@@ -8,7 +8,7 @@ import { addQueryArgs } from '@wordpress/url';
 import { useCheckoutUrl } from 'calypso/blocks/importer/hooks/use-checkout-url';
 import { LoadingEllipsis } from 'calypso/components/loading-ellipsis';
 import useAddHostingTrialMutation from 'calypso/data/hosting/use-add-hosting-trial-mutation';
-import useCheckEligibilityMigrationTrialPlan from 'calypso/data/plans/use-check-eligibility-migration-trial-plan';
+import { useCheckMigrationTrialPlanEligibility } from 'calypso/data/plans/use-check-eligibility-migration-trial-plan';
 import { useQuery } from 'calypso/landing/stepper/hooks/use-query';
 import { useSite } from 'calypso/landing/stepper/hooks/use-site';
 import { useSiteSlug } from 'calypso/landing/stepper/hooks/use-site-slug';
@@ -37,7 +37,7 @@ const MigrationTrialAcknowledgeInternal = function ( props: Props ) {
 	const { user, site, siteSlug, flowName, stepName, submit } = props;
 
 	const { data: migrationTrialEligibility, isLoading: isCheckingEligibility } =
-		useCheckEligibilityMigrationTrialPlan( site?.ID );
+		useCheckMigrationTrialPlanEligibility( site?.ID );
 	const isEligibleForTrialPlan = migrationTrialEligibility?.eligible;
 
 	const plan = getPlan( PLAN_BUSINESS );
